requirements.txt
flask
pyTelegramBotAPI
python-dotenv
requests
gunicorn
